requirements.txt
django
psycopg2
djangorestframework

black
pdbpp
